// src/types.ts
export interface StoreProduct {
  id: string;
  name: string;
  category: string;
  categoryLabel: string;
  mainCategory?: 'cars' | 'carpets' | 'furniture' | 'wash_offers' | 'camping_trips' | 'car_care' | 'ac' | string;
  subCategory?: string;
  subCategoryLabel?: string;
  price: number;
  originalPrice?: number;
  rating: number;
  reviewsCount: number;
  deliveryEstimate: string; // e.g. "1-3 أيام"
  deliveryCost?: number;
  freeDeliveryThreshold?: number;
  image: string;
  images?: string[];
  features?: { title: string; subtitle: string; iconType?: 'shield' | 'bolt' | 'devices' | 'quality' | 'truck' | 'star' | string }[];
  description: string;
  stock: number;
  isActive: boolean;
  isNew?: boolean;
}